+++ src/CGnuPlotUtil.h
@@ -9,12 +9,12 @@
 
 namespace CGnuPlotUtil {
   inline double avg(std::initializer_list<double> l) {
-    return std::accumulate(l.begin(), l.end(), 0.0)/l.size();
+    return std::accumulate(l.begin(), l.end(), 0.0)/double(l.size());
   }
 
   template<typename T>
   inline T avgT(std::initializer_list<T> l, const T &z=T(0)) {
-    T z1 = z; return std::accumulate(l.begin(), l.end(), z1)/l.size();
+    T z1 = z; return std::accumulate(l.begin(), l.end(), z1)/int(l.size());
   }
 
   inline double norm(double x, double low, double high) {
