+++ src/Expr/CExprIValue.h
@@ -11,13 +11,13 @@ class CExprIntegerValue : public CExprValueBase {
     return new CExprIntegerValue(integer_);
   }
 
-  bool getBooleanValue(bool        &b) const { b = (integer_ != 0); return true; }
-  bool getIntegerValue(long        &l) const { l = integer_       ; return true; }
-  bool getRealValue   (double      &r) const { r = integer_       ; return true; }
+  bool getBooleanValue(bool        &b) const { b = (integer_ != 0) ; return true; }
+  bool getIntegerValue(long        &l) const { l = integer_        ; return true; }
+  bool getRealValue   (double      &r) const { r = double(integer_); return true; }
   bool getStringValue (std::string &s) const;
 
   bool getComplexValue(std::complex<double> &c) const {
-    c = std::complex<double>(integer_, 0.0); return true;
+    c = std::complex<double>(double(integer_), 0.0); return true;
   }
 
   void setIntegerValue(long l) { integer_ = l; }
